feat: save theme preference to local storage and apply theme on load
